# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "GraphSpec": {
        "additionalProperties": false,
        "description": "The canonical specification for a pipeline execution graph.\n\nThis is the top-level model representing a complete run graph.\nIt is versioned and designed for JSON serialization.",
        "properties": {
          "concept_registry": {
            "additionalProperties": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Concept Registry",
            "type": "object"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "edges": {
            "items": {
              "$ref": "#/components/schemas/EdgeSpec"
            },
            "title": "Edges",
            "type": "array"
          },
          "graph_id": {
            "title": "Graph Id",
            "type": "string"
          },
          "meta": {
            "additionalProperties": true,
            "title": "Meta",
            "type": "object"
          },
          "nodes": {
            "items": {
              "$ref": "#/components/schemas/NodeSpec"
            },
            "title": "Nodes",
            "type": "array"
          },
          "pipe_registry": {
            "additionalProperties": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Pipe Registry",
            "type": "object"
          },
          "pipeline_ref": {
            "$ref": "#/components/schemas/PipelineRef"
          }
        },
        "required": [
          "graph_id",
          "created_at"
        ],
        "title": "GraphSpec",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LeanRunStatus": {
        "description": "Lifecycle status of a runner-tracked run. A lean subset of the platform's status set.",
        "enum": [
          "PENDING",
          "RUNNING",
          "COMPLETED",
          "FAILED"
        ],
        "title": "LeanRunStatus",
        "type": "string"
      },
...
      "PipeOutput": {
        "additionalProperties": false,
        "properties": {
          "graph_assembly_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Graph Assembly Error"
          },
          "graph_spec": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GraphSpec"
              },
              {
                "type": "null"
              }
            ]
          },
          "pipeline_run_id": {
            "default": "untitled",
            "title": "Pipeline Run Id",
            "type": "string"
          },
          "working_memory": {
            "$ref": "#/components/schemas/WorkingMemory"
          },
          "working_memory_raw": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Working Memory Raw"
          }
        },
        "title": "PipeOutput",
        "type": "object"
      },
...
      "PipelexPipelineExecuteResponse": {
        "properties": {
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "main_stuff_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Main Stuff Name"
          },
          "pipe_output": {
            "$ref": "#/components/schemas/PipeOutput"
          },
          "pipeline_run_id": {
            "title": "Pipeline Run Id",
            "type": "string"
          },
          "pipeline_state": {
            "$ref": "#/components/schemas/PipelineState"
          }
        },
        "required": [
          "pipeline_run_id",
          "created_at",
          "pipeline_state",
          "pipe_output"
        ],
        "title": "PipelexPipelineExecuteResponse",
        "type": "object"
      },
...
      "PipelineState": {
        "enum": [
          "RUNNING",
          "COMPLETED",
          "FAILED",
          "CANCELLED",
          "ERROR",
          "STARTED"
        ],
        "title": "PipelineState",
        "type": "string"
      },
...
      "ResolveStorageUrlRequest": {
        "additionalProperties": false,
        "properties": {
          "uri": {
            "description": "pipelex-storage:// URI to resolve",
            "maxLength": 512,
            "title": "Uri",
            "type": "string"
          }
        },
        "required": [
          "uri"
        ],
        "title": "ResolveStorageUrlRequest",
        "type": "object"
      },
      "ResolveStorageUrlResponse": {
        "properties": {
          "content_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Content type guessed from the URI extension",
            "title": "Content Type"
          },
          "expires_at": {
            "description": "UTC expiry timestamp of the presigned URL",
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "url": {
            "description": "Short-lived presigned URL, fetchable directly by the browser",
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "url",
          "expires_at"
        ],
        "title": "ResolveStorageUrlResponse",
        "type": "object"
      },
...
      "RunStatusResponse": {
        "properties": {
          "created_at": {
            "description": "ISO-8601 creation timestamp.",
            "title": "Created At",
            "type": "string"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "ISO-8601 completion timestamp, once terminal.",
            "title": "Finished At"
          },
          "pipeline_run_id": {
            "description": "The run id.",
            "title": "Pipeline Run Id",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/LeanRunStatus",
            "description": "Current lifecycle status."
          }
        },
        "required": [
          "pipeline_run_id",
          "status",
          "created_at"
        ],
        "title": "RunStatusResponse",
        "type": "object"
      },
...
      "UploadRequest": {
        "additionalProperties": false,
        "properties": {
          "content_type": {
            "anyOf": [
              {
                "maxLength": 255,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "MIME type (e.g. 'application/pdf')",
            "title": "Content Type"
          },
          "data": {
            "description": "File content as base64-encoded string (max 50 MiB decoded)",
            "maxLength": 69905071,
            "title": "Data",
            "type": "string"
          },
          "filename": {
            "description": "Original filename with extension (e.g. 'resume.pdf')",
            "maxLength": 512,
            "title": "Filename",
            "type": "string"
          }
        },
        "required": [
          "filename",
          "data"
        ],
        "title": "UploadRequest",
        "type": "object"
      },
      "UploadResponse": {
        "properties": {
          "filename": {
            "description": "Original filename",
            "title": "Filename",
            "type": "string"
          },
          "uri": {
            "description": "pipelex-storage:// URI for the uploaded file",
            "title": "Uri",
            "type": "string"
          }
        },
        "required": [
          "uri",
          "filename"
        ],
        "title": "UploadResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "WorkingMemory": {
        "additionalProperties": false,
        "properties": {
          "aliases": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Aliases",
            "type": "object"
          },
          "root": {
            "additionalProperties": {
              "$ref": "#/components/schemas/Stuff"
            },
            "title": "Root",
            "type": "object"
          }
        },
        "title": "WorkingMemory",
        "type": "object"
      }
    },
    "securitySchemes": {
      "BearerAuth": {
        "description": "Bearer token: the `API_KEY` (when `AUTH_MODE=api_key`) or a signed JWT (when `AUTH_MODE=jwt`). Omit when `AUTH_MODE=none`.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {
      "name": "Pipelex",
      "url": "https://docs.pipelex.com/"
    },
    "description": "The **Pipelex Runner API** is the open-source, self-hostable engine that executes\nPipelex pipelines (MTHDS bundles) over HTTP. It is **stateless**: it needs no\ndatabase, queue, or object store to run pipelines \u2014 boot the Docker image, give it\nan inference backend, and call `/api/v1/pipeline/execute`.\n\n### What you must provide\n- **Inference backend** \u2014 by default a `PIPELEX_GATEWAY_API_KEY` (the Pipelex Gateway\n  routes to every supported model with one credential). To call providers directly\n  (OpenAI, Anthropic, Bedrock, Vertex, \u2026) you reconfigure the mounted `.pipelex/`\n  config \u2014 not just an env var. See the Configuration docs.\n- **Auth mode** (`AUTH_MODE`) \u2014 `none` (default, local), `api_key` (`API_KEY`), or\n  `jwt` (`JWT_SECRET_KEY`). The caller is identified by the credential; the runner\n  never accepts a user id as a request argument.\n\n### What is NOT in this runner\nDurable run storage and by-id polling (`start`/`status`/`result`/`poll`), user-scoped\nfile storage (`/upload`, `/resolve-storage-url`), org/billing \u2014 these are **Pipelex\nPlatform** features. The open-source runner is execution-only; run a pipeline and get\nits output back synchronously via `/pipeline/execute`.\n",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/license/mit"
    },
...
    "version": "0.1.2"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/pipeline/execute": {
      "post": {
        "description": "Runs the pipeline and returns its full output in the response. This is the core self-hosted primitive \u2014 it has no time limit beyond what your own client/proxy allows. (Behind the hosted Pipelex gateway a ~30s synchronous cap applies; off-gateway it does not, but your own reverse proxy's read/idle timeout still does.)",
        "operationId": "execute_api_v1_pipeline_execute_post",
        "requestBody": {
          "content": {
            "application/json": {
              "example": {
                "inputs": {
                  "text": "Pipelex turns plain-language pipeline definitions into reproducible AI workflows."
                },
                "mthds_contents": [
                  "domain = \"hello\"\nmain_pipe = \"summarize\"\n\n[pipe.summarize]\ntype = \"PipeLLM\"\ndescription = \"Summarize the input text in one sentence\"\ninputs = { text = \"Text\" }\noutput = \"Text\"\nprompt = \"Summarize in one sentence:\\n@text\"\n"
                ],
                "pipe_code": "summarize"
              },
              "schema": {
                "properties": {
                  "dynamic_output_concept_ref": {
                    "description": "Optional concept ref to coerce the output to.",
                    "type": "string"
                  },
                  "inputs": {
                    "additionalProperties": true,
                    "description": "Pipeline inputs keyed by name. Values may be plain text, structured objects, or Document/Image references.",
                    "type": "object"
                  },
                  "mthds_contents": {
                    "description": "Inline MTHDS bundle(s) defining the pipe(s). Always an array, even for one file.",
                    "items": {
                      "type": "string"
                    },
                    "type": "array"
                  },
                  "output_multiplicity": {
                    "description": "Optional output multiplicity override (bool or int)."
                  },
                  "output_name": {
                    "description": "Optional name of the output stuff to return.",
                    "type": "string"
                  },
                  "pipe_code": {
                    "description": "Code of the pipe to run. Provide this and/or `mthds_contents`.",
                    "example": "summarize",
                    "type": "string"
                  }
                },
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PipelexPipelineExecuteResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Run a pipeline and wait for the result (blocking)",
        "tags": [
          "pipeline"
        ]
      }
    },
...
    "/api/v1/resolve-storage-url": {
      "post": {
        "description": "**Pipelex Platform feature \u2014 not part of the open-source self-host standard.** Requires `AUTH_MODE=jwt` (a per-user identity). Only needed for `pipelex-storage://` URIs produced by `/upload`; self-hosters using public URLs never need this.",
        "operationId": "resolve_storage_url_api_v1_resolve_storage_url_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolveStorageUrlRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResolveStorageUrlResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Resolve a pipelex-storage:// URI to a presigned URL (Pipelex Platform feature)",
        "tags": [
          "storage"
        ]
      }
    },
...
    "/api/v1/runs/by-id/{pipeline_run_id}": {
      "get": {
        "operationId": "get_run_api_v1_runs_by_id__pipeline_run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "pipeline_run_id",
            "required": true,
            "schema": {
              "title": "Pipeline Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get a run's status by id",
        "tags": [
          "runs"
        ]
      }
    },
    "/api/v1/runs/by-id/{pipeline_run_id}/result": {
      "get": {
        "description": "Single-shot result lookup. **202** while the run is pending/running (retry after `Retry-After` seconds); **200** with the execute response when COMPLETED; **409** when the run reached a terminal failure.",
        "operationId": "get_run_result_api_v1_runs_by_id__pipeline_run_id__result_get",
        "parameters": [
          {
            "in": "path",
            "name": "pipeline_run_id",
            "required": true,
            "schema": {
              "title": "Pipeline Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get a run's result by id (202 running / 200 completed / 409 failed)",
        "tags": [
          "runs"
        ]
      }
    },
    "/api/v1/upload": {
      "post": {
        "description": "**Pipelex Platform feature \u2014 not part of the open-source self-host standard.** Requires `AUTH_MODE=jwt` (a per-user identity to scope storage under). Self-hosters should pass a public HTTP(S) URL or a base64 data URL directly as a `Document`/`Image` input instead of uploading.",
        "operationId": "upload_file_api_v1_upload_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UploadRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upload a user-scoped file (Pipelex Platform feature)",
        "tags": [
          "uploader"
        ]
      }
    },
...
  },
  "security": [
    {},
    {
      "BearerAuth": []
    }
  ],
  "servers": [
...
}